+++ src/LiveCharts.Core/Charts/PolarChartModel.cs
@@ -58,7 +58,7 @@ namespace LiveCharts.Core.Charts
         }
 
         /// <inheritdoc />
-        public override float ScaleFromUi(float pixelsValue, Plane plane, float[] sizeVector = null)
+        public override double ScaleFromUi(float pixelsValue, Plane plane, float[] sizeVector = null)
         {
             throw new System.NotImplementedException();
         }
